backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
celery==5.2.7
redis==4.5.4
Werkzeug==2.2.3
python-dotenv==1.0.0
crewai==0.11.0
langchain==0.1.0
openai<2.00
flask-socketio==5.3.2
python-engineio==4.3.4
python-socketio==5.7.2 
PyPDF2
python-magic
pytesseract
Pillow
pymongo==4.6.1